src/requirements-linux.txt
absl-py==0.12.0
altgraph==0.17
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
ConfigObject==1.2.2
djitellopy2==2.3
flatbuffers==1.12
future==0.18.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
pefile==2019.4.18
Pillow==8.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python3-pygame
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pywin32-ctypes==0.2.0
recordtype==1.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
